Deflake akamai purger integration testing
